BaseProduct moved to folder productmodule.management
