+++ src/main/java/dk/sdu/se_f22/productmodule/infrastructure/domain/ProductInfSearch.java
@@ -1,6 +1,6 @@
 package dk.sdu.se_f22.productmodule.infrastructure.domain;
 
-import dk.sdu.se_f22.sharedlibrary.models.BaseProduct;
+import dk.sdu.se_f22.productmodule.management.BaseProduct;
 
 import java.util.List;
