+++ src/main/java/dk/sdu/se_f22/productmodule/management/IProductManager.java
@@ -1,7 +1,5 @@
 package dk.sdu.se_f22.productmodule.management;
 
-import dk.sdu.se_f22.sharedlibrary.models.BaseProduct;
-
 import java.util.ArrayList;
 
 public interface IProductManager {
