+++ src/test/java/dk/sdu/se_f22/productmodule/management/BaseProductTest.java
@@ -1,7 +1,5 @@
 package dk.sdu.se_f22.productmodule.management;
 
-import dk.sdu.se_f22.sharedlibrary.models.BaseProduct;
-
 import org.junit.jupiter.api.*;
 import java.io.IOException;
 import java.util.ArrayList;
